fix: send idempotency keys
