+++ services/ctl-api/internal/pkg/queue/client/await_signal.go
@@ -7,10 +7,10 @@ import (
 
 	"github.com/pkg/errors"
 
-	"go.temporal.io/sdk/activity"
 	tclient "go.temporal.io/sdk/client"
 	"go.temporal.io/sdk/temporal"
 
+	"github.com/nuonco/nuon/pkg/temporal/heartbeat"
 	"github.com/nuonco/nuon/services/ctl-api/internal/app"
 	dbgenerics "github.com/nuonco/nuon/services/ctl-api/internal/pkg/db/generics"
 	"github.com/nuonco/nuon/services/ctl-api/internal/pkg/queue/handler"
@@ -36,68 +36,53 @@ func (c *Client) AwaitSignal(ctx context.Context, queueSignalID string) (*handle
 		return &handler.FinishedResponse{}, nil
 	}
 
-	// Heartbeat so Temporal knows this activity is alive during the blocking update call.
-	done := make(chan struct{})
-	defer close(done)
-	go func() {
-		ticker := time.NewTicker(3 * time.Second)
-		defer ticker.Stop()
-		for {
-			select {
-			case <-done:
-				return
-			case <-ctx.Done():
-				return
-			case <-ticker.C:
-				activity.RecordHeartbeat(ctx, nil)
+	return heartbeat.WithHeartbeat(ctx, 3*time.Second, func(ctx context.Context) (*handler.FinishedResponse, error) {
+		rawResp, err := c.tClient.UpdateWorkflowInNamespace(ctx, q.Workflow.Namespace, tclient.UpdateWorkflowOptions{
+			UpdateID:     queueSignalID + "-finished",
+			WorkflowID:   q.Workflow.ID,
+			UpdateName:   handler.FinishedHandlerName,
+			WaitForStage: tclient.WorkflowUpdateStageCompleted,
+			Args: []any{
+				handler.FinishedRequest{},
+			},
+		})
+		if err != nil {
+			// Workflow may have been slept/terminated between our DB check and now.
+			// Re-check DB status to confirm.
+			fresh, dbErr := c.getQueueSignal(ctx, queueSignalID)
+			if dbErr != nil {
+				return nil, errors.Wrap(dbErr, "unable to get queue signal from db")
 			}
+			if isTerminalStatus(fresh.Status.Status) {
+				if fresh.Status.Status == app.StatusError {
+					return nil, temporal.NewNonRetryableApplicationError(
+						fmt.Sprintf("signal execution failed with status: %s", fresh.Status.Status),
+						"SIGNAL_FAILED", nil)
+				}
+				return &handler.FinishedResponse{}, nil
+			}
+			return nil, errors.Wrap(err, "unable to call finished handler")
 		}
-	}()
 
-	rawResp, err := c.tClient.UpdateWorkflowInNamespace(ctx, q.Workflow.Namespace, tclient.UpdateWorkflowOptions{
-		WorkflowID:   q.Workflow.ID,
-		UpdateName:   handler.FinishedHandlerName,
-		WaitForStage: tclient.WorkflowUpdateStageCompleted,
-		Args: []any{
-			handler.FinishedRequest{},
-		},
+		var resp handler.FinishedResponse
+		if err := rawResp.Get(ctx, &resp); err != nil {
+			return nil, errors.Wrap(err, "unable get response")
+		}
+
+		// Re-check DB status after workflow update completes, since the finishedHandler
+		// returns FinishedResponse{} regardless of whether signal execution succeeded or failed.
+		fresh, err := c.getQueueSignal(ctx, queueSignalID)
+		if err != nil {
+			return nil, errors.Wrap(err, "unable to verify signal status after completion")
+		}
+		if fresh.Status.Status == app.StatusError {
+			return nil, temporal.NewNonRetryableApplicationError(
+				fmt.Sprintf("signal execution failed with status: %s", fresh.Status.Status),
+				"SIGNAL_FAILED", nil)
+		}
+
+		return &resp, nil
 	})
-	if err != nil {
-		// Workflow may have been slept/terminated between our DB check and now.
-		// Re-check DB status to confirm.
-		fresh, dbErr := c.getQueueSignal(ctx, queueSignalID)
-		if dbErr != nil {
-			return nil, errors.Wrap(dbErr, "unable to get queue signal from db")
-		}
-		if isTerminalStatus(fresh.Status.Status) {
-			if fresh.Status.Status == app.StatusError {
-				return nil, temporal.NewNonRetryableApplicationError(
-					fmt.Sprintf("signal execution failed with status: %s", fresh.Status.Status),
-					"SIGNAL_FAILED", nil)
-			}
-			return &handler.FinishedResponse{}, nil
-		}
-		return nil, errors.Wrap(err, "unable to call finished handler")
-	}
-
-	var resp handler.FinishedResponse
-	if err := rawResp.Get(ctx, &resp); err != nil {
-		return nil, errors.Wrap(err, "unable get response")
-	}
-
-	// Re-check DB status after workflow update completes, since the finishedHandler
-	// returns FinishedResponse{} regardless of whether signal execution succeeded or failed.
-	fresh, err := c.getQueueSignal(ctx, queueSignalID)
-	if err != nil {
-		return nil, errors.Wrap(err, "unable to verify signal status after completion")
-	}
-	if fresh.Status.Status == app.StatusError {
-		return nil, temporal.NewNonRetryableApplicationError(
-			fmt.Sprintf("signal execution failed with status: %s", fresh.Status.Status),
-			"SIGNAL_FAILED", nil)
-	}
-
-	return &resp, nil
 }
 
 func (c *Client) getQueueSignal(ctx context.Context, id string) (*app.QueueSignal, error) {
